+++ Gruntfile.js
@@ -1,0 +1,2 @@
+require('./register')
+exports = module.exports = require('./project/Gruntfile.ts')