+++ register.js
@@ -1,0 +1,2 @@
+
+require("@babel/register")({ extensions: [ '.js', '.ts', '.tsx' ], ...require('./.babelrc.js') })