+++ src/www/vendor.d.ts
@@ -1,0 +1,3 @@
+
+declare const ReactDOM: typeof import('react-dom')
+declare const React: typeof import('react')