+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 
+# Development
+## 0.0.2
+`+` New grunt tasks (TS, SASS, Babel)  
+`+` Create entire React app with esmodules  
+`+` Create entire back-end server
+`*` Update a bunch of workspace configs  
+`-` Removed namespacing/workspaces on packages  
+
 ## 0.0.1
 `+` New project stuff, yay!